Add script "tempfix:dev" to run for development
New script uses --mode=production webpack option. The previous "dev"
script used --mode=development and had issues running on the browser.

Error: "t.isBuffer is not a function"

Because of this error, the Vue instance wouldn't show up when running
the "dev" script. The new "tempfix:dev" script allows for the Vue
instance to show up on the browser and to develop.